Fixing the array that is getting render and restarting user input after submit to an empty string
The array that was being used to render the user answers was returning null if there weren't any previous answers in the localStorage, stopping the map from loading.
Restarting userGuess to "", because it had always saved the previous answer if the user didn't type anything it would use the previous answer for submitting.